Separate stampy-search library
