+++ .gitignore
@@ -9,3 +9,6 @@ node_modules
 
 .DS_store
 wrangler.toml
+
+stampy-search/dist
+stampy-search/example/stampySearch.min.js
