+++ stampy-search/package.json
@@ -1,0 +1,29 @@
+{
+    "name": "stampy-search",
+    "version": "1.0.0",
+    "main": "dist/index.js",
+    "types": "dist/index.d.ts",
+    "scripts": {
+        "build": "tsc",
+        "build:prod": "NODE_ENV=production webpack --mode=production",
+        "eslint": "eslint --ignore-pattern .gitignore \"**/*.ts*\"",
+        "prettier": "prettier --check --ignore-path .gitignore \"**/*.{ts*,js,css}\"",
+        "lint": "tsc && npm run prettier && npm run eslint",
+        "eslint:fix": "eslint --fix --ignore-pattern .gitignore \"**/*.ts*\"",
+        "prettier:fix": "prettier --write --ignore-path .gitignore \"**/*.{ts*,js,css,md}\"",
+        "lint:fix": "tsc && npm run prettier:fix && npm run eslint:fix",
+        "example": "NODE_ENV=production webpack --mode=production && cp dist/stampySearch.min.js example && node example/server.js",
+        "test": "jest"
+    },
+    "devDependencies": {
+        "@babel/preset-react": "^7.0.0",
+        "@types/jest": "^29.4.0",
+        "babel-loader": "^8.0.0",
+        "jest": "^29.4.2",
+        "ts-jest": "^29.0.5",
+        "ts-loader": "^8.4.0",
+        "typescript": "^4.9.5",
+        "webpack": "^5.88.2",
+        "webpack-cli": "^5.1.4"
+    }
+}
