+++ stampy-search/src/index.ts
@@ -1,0 +1,2 @@
+export type {Question, SearchResult, WorkerMessage} from './search'
+export {searchLive, searchUnpublished, setupSearch} from './search'
