+++ stampy-search/tsconfig.json
@@ -1,0 +1,15 @@
+{
+    "compilerOptions": {
+        "module": "esnext",
+        "target": "es2015",
+        "declaration": true,
+        "outDir": "./dist",
+        "moduleResolution": "node"
+    },
+    "include": [
+        "./src/**/*"
+    ],
+    "exclude": [
+        "**/tests/**"
+    ]
+}
